+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes to Polarvargen
 
+### 0.2.0 (June 30, 2016)
+* Hungry for Coffee?
+* Update test, now check all entry points
+* Moved entry points from configuration to polarvargen.json
+* Add Gulp
+* Gulp use watch that you can use livereload at your browser
+
 ### 0.1.0 (May 22, 2016)
 * Add SCSS, LESS, Stylus preproc
 * Add some dubious test